update upload_image to working with front end
