+++ web/src/pages/api/images/upload_image.ts
@@ -1,3 +1,3 @@
 // NOTE: use src/api when you need to make gql request using codegen
-import { handleImageUpload } from "@/api/images/upload_image";
+import { handleImageUpload } from "@/api/images/upload_image_url";
 export default handleImageUpload;
